% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{AssignAttr}
\alias{AssignAttr}
\title{AssignAttr}
\usage{
AssignAttr(x, attr, value)
}
\arguments{
\item{x}{object}

\item{attr}{name of attribute (character)}

\item{value}{value of attribute}
}
\value{
The same object with attribute set to desired value.
}
\description{
Applies arbitrary setter functions to an object within a local scope and returns the object with attributes set.
}
\examples{

a <- cbind(1:2)
colnames(a) <- "numbers"
print(a)

b <- AssignAttr(cbind(1:2), "colnames", "numbers")

}
