% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartesianprod.R
\name{CartesianProd}
\alias{CartesianProd}
\title{CartesianProd}
\usage{
CartesianProd(..., stringsAsFactors = getOption("stringsAsFactors"))
}
\arguments{
\item{...}{arbitrary arguments passed to expand.grid}

\item{stringsAsFactors}{default taken from getOption("stringsAsFactors")}
}
\value{
A data frame.
}
\description{
Cartesian product, based on \code{expand.grid} but with row-major expansion (rather than column-major convention of expand.grid).
}
