% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopyExampleFiles.R
\name{CopyPkgFiles}
\alias{CopyPkgFiles}
\alias{SyncPkgFiles}
\title{CopyPkgFiles, SyncPkgFiles}
\usage{
CopyPkgFiles(path = ".", addargs, files, package)

SyncPkgFiles(path = ".", addargs = "", files, package)
}
\arguments{
\item{path}{[character] Location to copy files. If present directory (\code{"."}) or directory exists, only contents are copied. Otherwise, a directory of this name is created and contents are copied there.}

\item{addargs}{[character] Arguments to rsync for \code{SyncPkgFiles}. Unused by \code{CopyPkgFiles}.}

\item{files}{[character] Name of package files.}

\item{package}{[character] Name of package.}
}
\value{
Side effect of copying example files to \code{path}.
}
\description{
Copy or sync package files (e.g., from \code{inst/}) to desired location.
}
\section{Functions}{
\itemize{
\item \code{SyncPkgFiles}: Sync files
}}

