% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destructure.R
\docType{data}
\name{Destructure}
\alias{Destructure}
\title{Destructure, `[<-.result`}
\format{An object of class \code{result} of length 1.}
\usage{
Destructure
}
\arguments{
\item{x}{function first argument}

\item{...}{function additional argument}

\item{value}{values to be assigned in parent frame}
}
\value{
produces side effect of assignment of multiple values in the parent frame
}
\description{
Mimick behavior of destructuring bind (so named in Lisp). Also called  sequence unpacking in Python, and implemented by \code{deal} in MATLAB.
This definition is taken directly from G. Grothendieck's implementation of \code{List} as posted in R-help[1].
}
\examples{

## As provided by G. Grothendieck in R-help:
Destructure[QR,,QRaux]  <- qr(c(1,1:3,3:1))
Destructure[,Green,Blue]  <- col2rgb("aquamarine")
}
\references{
https://stat.ethz.ch/pipermail/r-help/2004-June/053343.html
}
\keyword{datasets}

