% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac2num.R
\name{Fac2Num}
\alias{Fac2Num}
\title{Fac2Num}
\usage{
Fac2Num(x, class = "numeric")
}
\arguments{
\item{x}{factor vector}

\item{class}{\code{"numeric"} or \code{"integer"}}
}
\value{
A numeric vector.
}
\description{
Factor to numeric
}
