% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.R
\name{Flip}
\alias{Flip}
\title{Flip}
\usage{
Flip(fn, pos)
}
\arguments{
\item{fn}{a function}

\item{pos}{position number or argument name}
}
\value{
A function in which the position of arguments is ordered.
}
\description{
Rearranges order of named function arguments (excludes dotted argument list). By default, reverses order of named function arguments . If \code{pos} argument is specified, these arguments are arranged to come first. Useful with *apply() functions.
}
\examples{

Foo <- function(x,y,z)
  x / y + z

Flip(Foo)
Flip(Foo,"z")
Flip(Foo,2:3)
Flip(Foo,c("y","z"))
}
