% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemgetter.R
\name{Itemgetter}
\alias{Itemgetter}
\title{Itemgetter}
\usage{
Itemgetter(...)
}
\arguments{
\item{...}{arbitrary indices passed to `[[` or `[`}
}
\value{
A function.
}
\description{
Inspired by python's \code{operator.itemgetter}. Mostly useful for use with *apply functions.
}
\examples{

List <- list(1, 2:3)
Df <- data.frame(x=1, y=2:3)
Mat <- data.matrix(Df)

ix <- Itemgetter(2)
ix(List)
ix(Df)
ix(Mat)

ix <- Itemgetter(1:2)
ix(List)
ix(Df)
ix(Mat)

ix <- Itemgetter(2, 2)
ix(Df)
ix(Mat)

ix <- Itemgetter(2, 1:2)
ix(Df)
ix(Mat)

ix <- Itemgetter(2, "y")
ix(Df)
ix(Mat)

ix <- Itemgetter(2, TRUE)
ix(Df)
ix(Mat)


ix <- Itemgetter(2, 2, drop=FALSE)
ix(Df)
ix(Mat)

ix <- Itemgetter("y")
ix(Df)

Map(Itemgetter(2), list(List, Df, Mat))
}

