% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemgetter.R
\name{Let}
\alias{Let}
\title{Let}
\usage{
Let(...)
}
\arguments{
\item{...}{see example}
}
\value{
Evaluated expression, which is the last argument passed to \code{Let}.
}
\description{
Just a novelty function to define local variables for evaluation.
}
\examples{

x <- 1
z <- 2

## instead of writing
with(list(x=2, y=3), z*(x+y))

## we can write
Let(x=2, y=3, z*(x+y))
}
