% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiIndex.R
\name{MultiIndex}
\alias{MultiIndex}
\title{MultiIndex}
\usage{
MultiIndex(x, collapse = ", ", format = identity)
}
\arguments{
\item{x}{vector}

\item{collapse}{(optional) separator}

\item{format}{(optional) additional method for formatting (provided with MultiIndex is called through apply(), for instance)}
}
\value{
A character vector.
}
\description{
A kudgey way of constructing a multi-element key as a character vector.
}

