% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartlist.R
\name{NamedList}
\alias{NamedList}
\title{NamedList}
\usage{
NamedList(...)
}
\arguments{
\item{...}{list elements}
}
\value{
List with named elements.
}
\description{
Automatically label list elements if names are not provided
}
\examples{

a <- 3
out <- NamedList(a, b=1)
}

