% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiIndex.R
\name{ParseIndex2dataframe}
\alias{ParseIndex2dataframe}
\alias{ParseIndex2dataframe.default}
\title{ParseIndex2dataframe}
\usage{
ParseIndex2dataframe(x, ...)

\method{ParseIndex2dataframe}{default}(x, sep = ", ", names = NULL,
  type.convert = TRUE)
}
\arguments{
\item{x}{Vector of MultiIndex}

\item{sep}{(optional) separator}

\item{names}{(optional) names for data frame}

\item{convert}{(optional) \code{TRUE}/\code{FALSE}}
}
\value{
A character vector.
}
\description{
Create data frame from MultiIndex.
}
