% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{PopulateEnv}
\alias{PopulateEnv}
\title{PopulateEnv}
\usage{
PopulateEnv(env.name, path, new = FALSE, ...)
}
\arguments{
\item{env.name}{environment name}

\item{path}{file(s) or directory containing R files}

\item{new}{T/F recreate environment}

\item{...}{further arguments passed to \code{list.files}}
}
\value{
Side effect of populating environment with sourced file contents.
}
\description{
Populates environment with functions in a file or directory.
}
\details{
This function provides a means of keeping functions (and data) in separate namespaces. Creates and attaches named environment to search() path if it doesn't already exist.
}

