% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{PreserveAttr}
\alias{PreserveAttr}
\title{PreserveAttr}
\usage{
PreserveAttr(.which)
}
\arguments{
\item{.which}{name of attribute to preserve from first argument of returned function}
}
\value{
A new function which accepts the function to be applied as its first argument, and its arguments follow. The first of these arguments will be used to obtain the attribute to be copied to the returned object.
}
\description{
This function returns a function that retains the desired attribute of its first argument.
}
\examples{

label <- sprintf("V\%d",seq(3))

id.float <- setNames(seq(.1,.3,.1),label)

var.1 <- setNames(1:3,label)
attr(var.1,"id") <- id.float

var.2 <- setNames(4:6,label)
attr(var.2,"id") <- id.float

PreserveAttr("id")(rbind,var.1,var.2)
}
