% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ReadRDA}
\alias{ReadRDA}
\title{ReadRDA}
\usage{
ReadRDA(x)
}
\arguments{
\item{x}{RDA file}
}
\value{
List of (named) contents of RDA file.
}
\description{
Load RDA (.rda) file and return contents as a list.
}

