% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursion.R
\name{Recursenode}
\alias{Recursenode}
\title{Recursenode}
\usage{
Recursenode(x)
}
\arguments{
\item{x}{a node; possible list object}
}
\value{
This function returns either the node or the result of further recursion.
}
\description{
This function decides if element is a single-element list or can be further branched. Called by Recursetree().
}
\examples{

#See Recursetree()
}

