% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{Split}
\alias{Split}
\alias{Splitc}
\alias{Splitc.matrix}
\alias{Splitr}
\alias{Splitr.matrix}
\title{Splitr, Splitc}
\usage{
Splitr(x, ...)

\method{Splitr}{matrix}(x)

Splitc(x, ...)

\method{Splitc}{matrix}(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
\code{Splitr}: Split matrix by row; return list.

\code{Splitc}: Split matrix by column; return list..
}
\description{
Split matrix by row or column
}

