% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiIndex.R
\name{MultiIndex}
\alias{MultiIndex}
\alias{MultiIndex.data.frame}
\alias{MultiIndex.default}
\alias{MultiIndex.list}
\alias{MultiIndex.matrix}
\title{MultiIndex}
\usage{
MultiIndex(x, ...)

\method{MultiIndex}{default}(x, collapse = ", ", format = identity)

\method{MultiIndex}{matrix}(x, ...)

\method{MultiIndex}{data.frame}(x, ...)

\method{MultiIndex}{list}(x, ...)
}
\arguments{
\item{x}{vector}

\item{collapse}{(optional) separator}

\item{format}{(optional) additional method for formatting (provided with MultiIndex is called through apply(), for instance)}
}
\value{
A character vector.
}
\description{
A kudgey way of constructing a multi-element key as a character vector.
}

