% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursion.R
\name{Recursetree}
\alias{Recursetree}
\title{Recursetree}
\usage{
Recursetree(x)
}
\arguments{
\item{x}{list object}
}
\value{
Parse tree of a function.
}
\description{
This function recurses a nested list to return every element of a node as a list element. One application is to obtain the parse tree of a function. Calls Recursenode().
}
\examples{

Foo <- function(x,y) {
  x / y
}

parsetree <- Recursetree(as.list(body(Foo)))

treeaslist <- Recursetree(list(1:5))
}

