% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetIndex.R
\name{SetIndex}
\alias{SetIndex}
\alias{SetIndex.default}
\alias{SetIndex.tbl_df}
\alias{ResetIndex}
\alias{ResetIndex.data.frame}
\alias{ResetIndex.matrix}
\title{SetIndex, ResetIndex}
\usage{
SetIndex(x, ...)

\method{SetIndex}{default}(x, index = colnames(x)[1])

\method{SetIndex}{tbl_df}(x, ...)

ResetIndex(x, ...)

\method{ResetIndex}{data.frame}(x, index = "index")

\method{ResetIndex}{matrix}(x, ...)
}
\arguments{
\item{x}{[data.frame or character matrix] table}

\item{index}{[character] index number or name}
}
\value{
[data.frame or matrix]
}
\description{
Set index like python's DataFrame.set_index(,inplace=False) but works with data frames and character matrices
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Sets row name to value of \code{index} and removes \code{index} from matix/data frame.

\item \code{tbl_df}: Converts tibble object (which does not accept row names) to data frame before applying \code{SetIndex.default}.

\item \code{data.frame}: Reset index (like \code{dplyr::add_rownames} or pandas \code{DataFrame.reset_index(,inplace=False)}). Returns data frame with index in first position.

\item \code{matrix}: Call as.data.frame on matrix and then apply ResetIndex.data.frame.
}}

